Working version of an Aries inbound transport for AMQP.
