+++ deploy/compose/canis-didcomm-lb.yaml
@@ -5,4 +5,16 @@
 ###############################################################
 inbound:
   host: 0.0.0.0
-  port: 7979
+  httpPort: 3003
+  wsPort: 3001
+
+###############################################################
+#
+#  AMQP Server (rabbitMQ)
+#
+###############################################################
+amqp:
+  user: guest
+  password: guest
+  host: localhost
+  port: 5672
